refactor : divide logic at useCalculate hook
